Consolidate styling code in pixiStyles.js
