+++ modules/pixi/index.js
@@ -15,4 +15,4 @@ export { pixiKeepRight } from './pixiKeepRight.js';
 export { pixiImproveOSM } from './pixiImproveOSM.js';
 export { pixiOsmose } from './pixiOsmose.js';
 export { pixiLayers } from './layers.js';
-
+export { styleMatch } from './pixiStyles.js';
