+++ modules/pixi/pixiAreas.js
@@ -2,8 +2,9 @@ import * as PIXI from 'pixi.js';
 import geojsonRewind from '@mapbox/geojson-rewind';
 import { vecLength, geomGetSmallestSurroundingRectangle } from '@id-sdk/math';
 
-import { getPixiTagPatternKey } from './pixiHelpers';
 import { prefs } from '../core/preferences';
+import { getPixiTagPatternKey } from './pixiHelpers';
+import { styleMatch } from './pixiStyles';
 
 const PARTIALFILLWIDTH = 32;
 
@@ -220,8 +221,8 @@ if (isOuter && !feature.ssrdata && feature.polygons.length === 1) {
 
 
         // Determine style info
-        let color = feature.style.color || 0xaaaaaa;
-        let alpha = feature.style.alpha || 0.25;
+        let color = feature.style.fill.color || 0xaaaaaa;
+        let alpha = feature.style.fill.alpha || 0.3;
         let texture = feature.texture || PIXI.Texture.WHITE;  // WHITE turns off the texture
         let doPartialFill = (fillstyle === 'partial');
 
@@ -269,7 +270,7 @@ if (isOuter && !feature.ssrdata && feature.polygons.length === 1) {
           .clear()
           .lineStyle({
             alpha: 1,
-            width: feature.style.width || 2,
+            width: feature.style.fill.width || 2,
             color: color
           });
 
@@ -355,187 +356,4 @@ if (isOuter && !feature.ssrdata && feature.polygons.length === 1) {
   }
 
   return renderAreas;
-}
-
-
-const STYLES = {
-  red: {
-    width: 2,
-    color: 0xe06e5f,  // rgba(224, 110, 95)
-    alpha: 0.3
-  },
-  green: {
-    width: 2,
-    color: 0x8cd05f,  // rgb(140, 208, 95)
-    alpha: 0.3
-  },
-  blue: {
-    width: 2,
-    color: 0x77d4de,  // rgb(119, 211, 222)
-    alpha: 0.3
-  },
-  yellow: {
-    width: 2,
-    color: 0xffff94,  // rgb(255, 255, 148)
-    alpha: 0.25
-  },
-  gold: {
-    width: 2,
-    color: 0xc4be19,  // rgb(196, 189, 25)
-    alpha: 0.3
-  },
-  orange: {
-    width: 2,
-    color: 0xd6881a,  // rgb(214, 136, 26)
-    alpha: 0.3
-  },
-  pink: {
-    width: 2,
-    color: 0xe3a4f5,  // rgb(228, 164, 245)
-    alpha: 0.3
-  },
-  teal: {
-    width: 2,
-    color: 0x99e1aa,  // rgb(153, 225, 170)
-    alpha: 0.3
-  },
-  lightgreen: {
-    width: 2,
-    color: 0xbee83f,  // rgb(191, 232, 63)
-    alpha: 0.3
-  },
-  tan: {
-    width: 2,
-    color: 0xf5dcba,  // rgb(245, 220, 186)
-    alpha: 0.3
-  },
-  darkgray: {
-    width: 2,
-    color: 0x8c8c8c,  // rgb(140, 140, 140)
-    alpha: 0.5
-  },
-  lightgray: {
-    width: 2,
-    color: 0xaaaaaa,  // rgb(170, 170, 170)
-    alpha: 0.3
-  }
-};
-
-const TAGSTYLES = {
-  amenity: {
-    fountain: 'blue',
-    childcare: 'yellow',
-    kindergarten: 'yellow',
-    school: 'yellow',
-    college: 'yellow',
-    university: 'yellow',
-    research_institute: 'yellow',
-    parking: 'darkgray'
-  },
-  building: {
-    '*': 'red'
-  },
-  construction: {
-    '*': 'gold'
-  },
-  barrier: {
-    hedge: 'green'
-  },
-  golf: {
-    'green': 'lightgreen'
-  },
-  landuse: {
-    flowerbed: 'green',
-    forest: 'green',
-    grass: 'green',
-    recreation_ground: 'green',
-    village_green: 'green',
-    residential: 'gold',
-    retail: 'orange',
-    commercial: 'orange',
-    landfill: 'orange',
-    military: 'orange',
-    industrial: 'pink',
-    cemetery: 'lightgreen',
-    farmland: 'lightgreen',
-    meadow: 'lightgreen',
-    orchard: 'lightgreen',
-    vineyard: 'lightgreen',
-    farmyard: 'tan',
-    railway: 'darkgray',
-    quarry: 'darkgray'
-  },
-  leisure: {
-    garden: 'green',
-    golf_course: 'green',
-    nature_reserve: 'green',
-    park: 'green',
-    pitch: 'green',
-    track: 'yellow',
-    swimming_pool: 'blue'
-  },
-  man_made: {
-    adit: 'darkgray',
-    groyne: 'darkgray',
-    breakwater: 'darkgray'
-  },
-  military: {
-    '*': 'orange'
-  },
-  natural: {
-    bay: 'blue',
-    water: 'blue',
-    beach: 'yellow',
-    sand: 'yellow',
-    scrub: 'yellow',
-    wetland: 'teal',
-    bare_rock: 'darkgray',
-    cave_entrance: 'darkgray',
-    cliff: 'darkgray',
-    rock: 'darkgray',
-    scree: 'darkgray',
-    stone: 'darkgray',
-    shingle: 'darkgray',
-    glacier: 'lightgray',
-    '*': 'green'
-  },
-  power: {
-    'plant': 'pink'
-  },
-  sport: {
-    beachvolleyball: 'yellow',
-    baseball: 'yellow',
-    softball: 'yellow',
-    basketball: 'darkgray',
-    skateboard: 'darkgray'
-  },
-  waterway: {
-    dam: 'darkgray',
-    weir: 'darkgray'
-  }
-};
-
-
-export function styleMatch(tags) {
-  let style = STYLES.lightgray;  // default
-  let selectivity = 999;
-
-  for (const k in tags) {
-    const v = tags[k];
-    const group = TAGSTYLES[k];
-    if (!group || !v) continue;
-
-    // smaller groups are more selective
-    let groupsize = Object.keys(group).length;
-    let stylename = group[v];
-    if (!stylename) stylename = group['*'];  // fallback value
-
-    if (stylename && groupsize < selectivity) {
-      style = STYLES[stylename];
-      selectivity = groupsize;
-      if (selectivity === 1) break;  // no need to keep looking at tags
-    }
-  }
-
-  return style;
 }
